backend/requirements.txt
# Academic Grievance DSS - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Python-Java Bridge
JPype1==1.5.0

# LLM Integration
openai==1.10.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Configuration
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Logging
loguru==0.7.2

# JSON Processing
orjson==3.9.12

# Date/Time
python-dateutil==2.8.2

# CORS
python-cors==1.0.0
